requirements.txt
selenium
paho-mqtt